logoipsum/clientes/serializers/cliente_serializer.py
from rest_framework import serializers
from clientes.models.cliente import Cliente


class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = "__all__"
